Update FindTheBrain.py
Changing the model, the number of layers is a tuple and not a simple int. Mistake until now, only use 3 layers(default) with n nodes.
'adam' is not a better solver, but a quicker with the same results than 'lbfgs'
In order to evaluate the influence of each parameter, a design of experiment is set up.