Create FindModel.py
Export the best model and its output.

Based on 7 differents methods + 1 ensemble, it compare all of them based on F-measure. Every methods that arrive above 0.15 are stacked in a list and then tested on another set of data to determine the best of best.

Size of dataset is constantly increase and ratio is a random number between 0.0005 and 0.005